Passwort wird vor Änderung validiert! Bei Abgabe einer Bewertung muss ein Kommentar eingegeben werden!
Änderung des Preises eines Artikels in Double um Cent werte anzuzeigen!.
Funktioniert aber noch nicht richtig!!
Integerwerte werden im Netclient bei getAllAngebote() in Double
umgerechnet!

+++ HsKafeteria/src/de/app/hskafeteria/AngeboteListAdapter.java
@@ -47,7 +47,7 @@ public class AngeboteListAdapter extends BaseExpandableListAdapter {
   childItem.setText(angebot.getTitel().trim());
   
   TextView childItemPrice = (TextView) view.findViewById(R.id.angebot_childItem_price);
-  Integer price = angebot.getPreis();
+  Double price = angebot.getPreis();
   childItemPrice.setText(price.toString() + "€");
